+++ qdl_klayout_extension/core/__init__.py
@@ -4,9 +4,9 @@ Such functionalities range from creating Coordinates, to creating Layouts with c
 in repeated patterns. It serves to ease and automate the creation of complex shapes and patterns.
 """
 
-import coordinates
-import geometries
-import layout
-import patterns
-import shapes
+from coordinates import *
+from geometries import *
+from layout import *
+from patterns import *
+from shapes import *